Hard code fix for #61
I don't have the availability right now for a complete debug and fix of
this issue. Right now, I'm hard coding it by resetting the user badge
counter everytime there is a new badge redeem.

It's ugly, has poor performance and I'm very sorry but we need this to
run smoothly right now.